requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
bcrypt==3.1.6
blinker==1.4
boto3==1.9.142
botocore==1.12.142
cffi==1.12.2
Click==7.0
cryptography==2.6.1
docutils==0.14
Flask==1.0.2
Flask-Limiter==1.0.1
Flask-Mail==0.9.0
Flask-Markdown==0.3
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.4
limits==1.3
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.1.0
passlib==1.7.1
Pillow==5.4.1
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-slugify==2.0.1
s3transfer==0.2.0
six==1.12.0
SQLAlchemy==1.2.17
Unidecode==1.0.23
urllib3==1.24.3
Werkzeug==0.14.1
WTForms==2.2.1
